+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Bug Fixes
 
+- [730](https://github.com/persistenceOne/pstake-native/pull/730) Fix deposit validate.  
 - [728](https://github.com/persistenceOne/pstake-native/pull/728) Fix prevent users from liquid-staking funds by
   removing the Deposit entry.
 - [726](https://github.com/persistenceOne/pstake-native/pull/726) Fix minimal unbondings.